+++ test/request/jwt_request.test.js
@@ -37,7 +37,7 @@ describe('request parameter features', function () {
           client_id: 'client',
           response_type: 'code',
           redirect_uri: 'https://client.example.com/cb'
-        }, null, 'none').then(request => this.wrap({
+        }, null, 'none', { issuer: 'client', audience: this.provider.issuer }).then(request => this.wrap({
           agent: this.agent,
           route,
           verb,
@@ -65,7 +65,7 @@ describe('request parameter features', function () {
           response_type: 'code',
           redirect_uri: 'https://client.example.com/cb',
           claims: JSON.stringify({ id_token: { email: null } })
-        }, null, 'none').then(request => this.wrap({
+        }, null, 'none', { issuer: 'client', audience: this.provider.issuer }).then(request => this.wrap({
           agent: this.agent,
           route,
           verb,
@@ -93,7 +93,7 @@ describe('request parameter features', function () {
           response_type: 'code',
           redirect_uri: 'https://client.example.com/cb',
           claims: { id_token: { email: null } }
-        }, null, 'none').then(request => this.wrap({
+        }, null, 'none', { issuer: 'client', audience: this.provider.issuer }).then(request => this.wrap({
           agent: this.agent,
           route,
           verb,
@@ -123,7 +123,7 @@ describe('request parameter features', function () {
           client_id: 'client-with-HS-sig',
           response_type: 'code',
           redirect_uri: 'https://client.example.com/cb'
-        }, key, 'HS256').then(request => this.wrap({
+        }, key, 'HS256', { issuer: 'client-with-HS-sig', audience: this.provider.issuer }).then(request => this.wrap({
           agent: this.agent,
           route,
           verb,
@@ -154,7 +154,7 @@ describe('request parameter features', function () {
           response_type: 'code',
           request: 'request inception',
           redirect_uri: 'https://client.example.com/cb'
-        }, null, 'none').then(request => this.wrap({
+        }, null, 'none', { issuer: 'client', audience: this.provider.issuer }).then(request => this.wrap({
           agent: this.agent,
           route,
           verb,
@@ -183,7 +183,7 @@ describe('request parameter features', function () {
           response_type: 'code',
           request_uri: 'request uri inception',
           redirect_uri: 'https://client.example.com/cb'
-        }, null, 'none').then(request => this.wrap({
+        }, null, 'none', { issuer: 'client', audience: this.provider.issuer }).then(request => this.wrap({
           agent: this.agent,
           route,
           verb,
@@ -211,7 +211,7 @@ describe('request parameter features', function () {
           client_id: 'client',
           response_type: 'id_token',
           redirect_uri: 'https://client.example.com/cb'
-        }, null, 'none').then(request => this.wrap({
+        }, null, 'none', { issuer: 'client', audience: this.provider.issuer }).then(request => this.wrap({
           agent: this.agent,
           route,
           verb,
@@ -239,7 +239,7 @@ describe('request parameter features', function () {
           client_id: 'client2',
           response_type: 'code',
           redirect_uri: 'https://client.example.com/cb'
-        }, null, 'none').then(request => this.wrap({
+        }, null, 'none', { issuer: 'client2', audience: this.provider.issuer }).then(request => this.wrap({
           agent: this.agent,
           route,
           verb,
@@ -292,7 +292,7 @@ describe('request parameter features', function () {
           client_id: 'client-with-HS-sig',
           response_type: 'code',
           redirect_uri: 'https://client.example.com/cb'
-        }, null, 'none').then(request => this.wrap({
+        }, null, 'none', { issuer: 'client-with-HS-sig', audience: this.provider.issuer }).then(request => this.wrap({
           agent: this.agent,
           route,
           verb,
@@ -324,7 +324,7 @@ describe('request parameter features', function () {
           client_id: 'client',
           response_type: 'code',
           redirect_uri: 'https://client.example.com/cb'
-        }, key, 'HS256').then(request => this.wrap({
+        }, key, 'HS256', { issuer: 'client', audience: this.provider.issuer }).then(request => this.wrap({
           agent: this.agent,
           route,
           verb,
@@ -340,7 +340,7 @@ describe('request parameter features', function () {
           expect(spy.calledOnce).to.be.true;
           expect(spy.args[0][0]).to.have.property('message', 'invalid_request_object');
           expect(spy.args[0][0]).to.have.property('error_description').that.matches(
-            /could not validate request object signature/
+            /could not validate request object/
           );
         }));
       });
@@ -354,7 +354,7 @@ describe('request parameter features', function () {
           unrecognized: true,
           response_type: 'code',
           redirect_uri: 'https://client.example.com/cb'
-        }, key, 'HS256').then(request => this.wrap({
+        }, key, 'HS256', { issuer: 'client-with-HS-sig', audience: this.provider.issuer }).then(request => this.wrap({
           agent: this.agent,
           route,
           verb,